Added Monte Carlo and Zu Chongzhi algorithms. Made the code more efficient and fixed pow and sqrt in Math utilities

+++ src/main/java/tunudo/dotstudios/net/calculations/Calculation.java
@@ -4,6 +4,6 @@ import java.math.BigDecimal;
 
 public abstract class Calculation {
     public abstract BigDecimal calculate(PI.Algorithms algorithm, BigDecimal precision,long customIterations);
-    public abstract BigDecimal calculate(PI.Algorithms algorithm, BigDecimal precision);
+    public abstract BigDecimal calculate(PI.Algorithms algorithm, long precision);
     public abstract long iterations(long precision, PI.Algorithms algorithm);
 }
